Renaming + Recoding of Anisotropic Functions

+++ include/Marmot/MarmotMath.h
@@ -144,7 +144,7 @@ namespace Marmot {
        	Matrix3d DirectionCosLocalToGlobal(const Matrix3d& LocalCoordinateSystem);
 	Matrix3d OrthonormalCoordinateSystem(const Vector3d& normalVector);	
         double Polyfit(const Eigen::Matrix<double,4,1>& Xdata, const Eigen::Matrix<double,4,1>& Ydata, int& angle );
-	Eigen::MatrixXi DyadicProduct(const Eigen::VectorXd& Vector1, const Eigen::VectorXd& Vector2);
+	Eigen::MatrixXd DyadicProduct(const Eigen::VectorXd& Vector1, const Eigen::VectorXd& Vector2);
 
     } // namespace Math
 } // namespace Marmot
